requirements.txt
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.7
decorator==5.1.0
distlib==0.3.3
filelock==3.3.2
greenlet==1.1.2
idna==3.3
jedi==0.18.0
matplotlib-inline==0.1.3
msgpack==1.0.2
Naked==0.1.31
numpy==1.21.4
parso==0.8.2
pbr==5.7.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycodestyle==2.8.0
pycryptodome==3.11.0
Pygments==2.10.0
pynvim==0.4.3
PyYAML==6.0
requests==2.26.0
shellescape==3.8.1
six==1.16.0
stevedore==3.5.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
